#13
I removed remove the source links from the help pages and vignette.
I moved the contents of NEWS to NEWS.md.
I update the contents of inst/CITATION to include all authors, correct year, correct version number, and URL of the new website (where it will be after we push it).

Also, the examples for the effectsize functions had redundant examples, so I removed the redundancies.